Reorganized
Set HKSM as root folder; divided aspects of application, editor, data,
and panel; renamed IntJTextField to IntField
